refactor: split adt module into files
Tests are split into unit tests and integ tests.

I updated a few non-public stuffs into public. I also created
Image::new_from_path.

The failing unit test will be fixed separately:
adt::background::tests::svg_background_linear_gradient
